Restructure bootstrap sequence
The bootstrap environment is now made distinct from the environment that
hosts Conda Constructor, in order to minimize its package trail. This
enables fetching as many of the dependencies as possible as PyPI wheels.
From there, the install script can merely deploy inside of an already
existing Python VEnv.

Also, moved the configuration settings over to a distinct submakefile.
This facilitates changing the Python version the installer expects.